clarify indiv vs project efforts
Per issues #1 and #8, I revised the language to make it clear when a funding effort was led by an individual vs. a project. Some of these cases are somewhat vague (ex. when a project consists of only one person), but I did my best. The distinction matters more IMO when a project is very big and has separate governance (ex. Django). Feel free to open an issue if you see one that you think is marked incorrectly!