add freemium licenses to list
I recognize that freemium licenses (ex. Fair Source, BSL) are not actually open source, and I've made that explicit in the description. However, it seems they are trying to answer the same question outlined in this guide, so it seems that anybody interested in this topic should know that these types of licenses exist.